feat: clean up structure & naming
