+++ src/builder/types.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2022.
+ * Copyright (c) 2021-2025.
  * Author Peter Placzek (tada5hi)
  * For the full copyright and license information,
  * view the LICENSE file that was distributed with this source code.
@@ -7,9 +7,9 @@
 
 import type {
     FieldsBuildInput, FiltersBuildInput, PaginationBuildInput, RelationsBuildInput, SortBuildInput,
-} from '../../builder';
-import type { Parameter, URLParameter } from '../../constants';
-import type { ObjectLiteral, ObjectLiteralKeys } from '../../types';
+} from './index';
+import type { Parameter, URLParameter } from '../constants';
+import type { ObjectLiteral, ObjectLiteralKeys } from '../types';
 
 type BuildParametersInput<
     T extends ObjectLiteral = ObjectLiteral,
