+++ src/index.ts
@@ -5,7 +5,7 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './query';
+export * from './builder';
 export * from './errors';
 export * from './constants';
 export * from './parser';
