+++ src/parser/index.ts
@@ -7,4 +7,6 @@
 
 export * from './parameters';
 
+export * from './base';
 export * from './module';
+export * from './types';
