--- src/query/parser/index.ts
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2025.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-export * from './module';
-export * from './types';
